Rename some internal components
