+++ src/evaluator/linkedlist.js
