+++ src/evaluator/linkedlist.spec.js
