+++ src/view/convulator.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import {Buttons} from './button'
-import {Line} from './line'
+import {Display} from './display'
 
 export function Convulator ({state, dispatcher}) {
   return (
     <div>
-      <Line expression={state.tokens} result={state.result} />
+      <Display expression={state.tokens} result={state.result} />
       <Buttons dispatcher={dispatcher} />
     </div>
   )
