+++ src/view/token.spec.js
